rename zod schemas folder
