+++ lib/validate-req-body.js
@@ -1,4 +1,4 @@
-import { zodSchemas } from "../zod-schemas/zod-schemas.js";
+import { zodSchemas } from "../zod/zod-schemas.js";
 import { throwError } from "./throw-error.js";
 
 // Note: this function is used to validate req.body before any other actions takes place
